requirements.txt
Flask==3.1.0
flask_limiter==3.12
loguru==0.7.3
pandas==2.2.3
python-dotenv==1.0.1
SQLAlchemy==2.0.38
waitress==3.0.2
